finshing strategy
TODO: remember always to make a final class in which the code is not elaborated for the DEMO presentation. This is important for showing only the interested lines of code... OR: keep separated the 2 files.
From now on, going with all the other patterns and come back to the STRATEGY modification of this aspect later on.