end of the Observer implementation and demo
a new bug inserted in the document for the presentation of the demo in which a part remains with the white background.
missing patterns:
state
template method,
visitor
TODO: before new additions in the DB: duplicating it and accessing it with a string connection in the config file. This for the part of the creation of new patterns from the interface.